+++ android/app/lib/providers/settings.g.dart
@@ -1,0 +1,26 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'settings.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$settingsHash() => r'a7222c38a2c7826d4561debf5b275655a7810930';
+
+/// See also [Settings].
+@ProviderFor(Settings)
+final settingsProvider =
+    AutoDisposeAsyncNotifierProvider<Settings, SettingsValue>.internal(
+      Settings.new,
+      name: r'settingsProvider',
+      debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+          ? null
+          : _$settingsHash,
+      dependencies: null,
+      allTransitiveDependencies: null,
+    );
+
+typedef _$Settings = AutoDisposeAsyncNotifier<SettingsValue>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package